split cfg part 1
